Fix incorrect assert in sr_port/op_wttab.c if LENGTH deviceparameter of $PRINCIPAL is set to 0
Below is a test case that demonstrates the issue.

YDB>use $p:(length=0)
YDB>write ?1
%YDB-F-ASSERT, Assert failed in sr_port/op_wttab.c line 80 for expression (0 != iod->length || iod->dollar.y < iod->length)

iod->dollar.y should be compared against iod->length only if iod->length is non-zero.
This is because iod->dollar.y is not maintained if iod->length is 0.

But the assert does exactly the opposite check. The fix is to change the "0 !=" to a "0 ==".