[StaticScan] sr_port/coerce.c : Fix potential uninitialized use of variable "conv"
"conv" was being set only in case the opcode was OC_LIT and not in case old_op was OC_COMVAL
or OC_COMINT. But "conv" is being used inside the if block even in the latter case which is an
uninitialized use. In that case though, we are guaranteed ref->operand[0].oprclass is TRIP_REF due
to the assert done just a line above where "conv" is initialized. And so, it is okay to set "conv"
even if old_op is OC_FORCENUM. This avoids the uninitialized use.