Ensure $Y never becomes equal to LENGTH of the terminal device
Recently, an assert in op_wttab.c was fixed (commit SHA e190d040b2f25a6f361c79ebffa902841f40ee49)
so it does the right checks (that $y is always < LENGTH) that it was originally intended for.

In internal testing, the revised assert failed in the expect/D9F05002548 subtest as follows.

%YDB-F-ASSERT, Assert failed in sr_port/op_wttab.c line 83 for expression ((0 == iod->length) || (iod->dollar.y < iod->length))

This is easily reproducible. In a terminal with say 10 lines or so run "gde show -reg" and you
will see the assert. It is a longstanding code issue because what the above assert failure means
is that $Y can be equal to the device LENGTH in some cases. But that should never be possible
according to the documentation.

https://docs.yottadb.com/ProgrammersGuide/isv.html#y

$Y never equals or exceeds the value of the device LENGTH. Whenever it reaches the value equal
to the device LENGTH, it gets reset to zero (0).

In iott_write.c, there is a place where we forget to maintain this condition. That is now fixed
and the expect/D9F05002548 test passes fine after this change.