[#347] Issue error at start of ydbinstall if any one of needed utilities is not installed on the system (!351)
We had a failure with ydbinstall.sh on a CentOS 7.5 docker image while trying to install the utf8 directory.
This turned out to be due to the absence of the "file" utility which ydbinstall.sh/configure.gtc assumed.

This is now fixed by checking for the existence of "file" and various other utilities assumed by the
ydbinstall.sh/configure.gtc scripts at the start of ydbinstall.sh. And issuing an error and exiting
if at least one of the needed utilities is not found.

The check is done using the "command -v" command of the POSIX sh. This is the recommended approach for
checking the existence of a utility instead of the 'which' command. So existing 'which' usages have been
replaced too.