Fix occasional failures in concurrently running tests that use offset.csh (!352)
In internal testing, we had a failure in the v62000/gtm7824 subtest with the following diff.

> v62000_0_5/gtm7824/dse_cache_show.out
> %YDB-E-CLIERR, Unrecognized option : OFFSET, value expected but not found

The test does the following.

     36 $gtm_tools/offset.csh node_local gtm_main.c >&! offset.out

The contents of the file in the test output directory show the issue.

> cat offset.out
/tmp/__nars_offset__gtm_main.out: Text file busy.

This caused a hexoffset computation to result in a "" return instead of a numeric value
causing later problems with a DSE CACHE -SHOW command invoked in com/waitforOLIstart.csh.

The reason for the "Text file busy" error is because the offset.csh script creates an
executable in /tmp with just the userid and C program name in the executable file name.

So if two concurrently running tests under the same userid run offset.csh with the same
C program name (in this case gtm_main.c), both will try creating an executable in the
same file path resulting in one of them encountering the "Text file busy" error.

This collision is now avoided by using $$ (the PID of the parent process) in the file
name. That would ensure two concurrently running tests will write to different files.