[#350] Fix duplicate keystroke(s) printing at direct mode or READ prompt in case of concurrent job interrupts (!354)
The v52000/D9G12002636 subtest failed once in many runs with the following diff

33c33
<
YDB>[?1h=z[Cc[Co[Cn[Ct[Ci[Cn[Cu[Ce[C[?1l>
---
>
YDB>[?1h=z[Cc[Co[Cn[Ct[Cii[Cn[Cu[Ce[C[?1l>

The user entered "zcontinue" at the direct mode prompt. But it showed up on the terminal as "zcontiinue".
Notice the double i.

This is because a mupip intrpt is done in this test a lot of times while the user is entering the
keystrokes. And terminal echo is turned on in mdb_condition_handler.c while in the middle of the direct
mode read as part of job interrupt handling which means any keystrokes the user enters until the job
interrupt returns will be displayed on the terminal twice, once by the terminal because ECHO setting is
now ON and another by the dm_read/iott_* routine when it sees those keystrokes.

This is now fixed by treating ERR_JOBINTRRQST (the error that is thrown on a MUPIP INTRPT) as a special
error and not doing the RESETTERM in that case in mdb_condition_handler.