Fix occasional v61000/gtm7923 test failure in debug builds
Once in around 100 or more test runs, we see a failure with the following diff

> v61000_0_1/gtm7923/child_gtm7923.mje2
> %YDB-F-ASSERT, Assert failed in sr_port/patstr.c line 818 for expression (atom_map < MAX_PATTERN_ATOMS)

The gtm7923 subtest generates random patterns which explains why it fails only once in a while.
Nevertheless, the assert that fails is newly introduced in GT.M V6.3-003 and is verifying that
we do not overflow the min[]/max[] array.

Turns out we do not overflow it whenever we access it through min[count], max[count] since we
have checks to ensure "count" never gets more than MAX_PATTERN_ATOMS.

But we do not have such a check for "atom_map" which is usually the same as "count" but can be
different in case a DFA pattern compilation is in progress. GT.M V6.3-003 seems to have introduced
a check for "atom_map" against MAX_PATTERN_ATOMS (and an accompanying PATMAXLEN error return) in
a restricted block of code. Not clear why that was done.

It seems to make sense to me to have this check just before when "atom_map" is incremented. So
I have moved that logic to the appropriate place.

In pro builds, we have always seen a PATMAXLEN failure, never any symptoms of a buffer overflow.
This, across thousands of test runs. So it is almost impossible to encounter this in practice.
Since there is no user-visible symptom that is discernible yet, no separate issue is created for
this on gitlab. A merge request corresponding to this commit seems good enough.