[#352] Return error status from ydb_ci() call back to C caller instead of GTMASSERT2 fatal error
The issue was that in gvcst_put(), if a spanning node set occurs, we set the SSF_NORET_VIA_MUMTSTART
bit in frame_pointer->flags to indicate (just like we do for triggers) that we should not ever return
via MUM_TSTART macro from this frame_pointer. That is because this is an implicit TP transaction
started in this C function gvcst_put(). But we never reset this bit when the implicit TP is
committed. This is fortunately okay in case of a pure M application because if ever error_return()
is invoked (because we have an unhandled runtime error) we would rethrow the error in the parent
frame (ERR_REPEATERROR) or EXIT. The only exception is with call-ins in which case we would proceed
further to do a MUM_TSTART ("rethrow_needed" variable would be reset to FALSE in that case).

So if a call-in happens after a ydb_set_s() call that sets a spanning node, the SSF_NORET_VIA_MUMTSTART
bit will still be set in frame_pointer->flags. If this call-in has an unhandled error, error_return()
would invoke MUM_TSTART which will GTMASSERT2 because the SSF_NORET_VIA_MUMTSTART bit is set.

An example GTMASSERT2 error is the below.

%YDB-F-GTMASSERT2, YottaDB r998 Linux x86_64 - Assert failed sr_unix/error_return.c line 134
	for expression (!(SSF_NORET_VIA_MUMTSTART & frame_pointer->flags) || (0 != proc_act_type) || (SFF_ETRAP_ERR & frame_pointer->flags))

The fix is two-fold.

1) Fix gvcst_put() to cleanup after itself by clearing the SSF_NORET_VIA_MUMTSTART bit when the
   implicit TP transaction is committed.

2) In case an error occurs before the op_tcommit() happens in gvcst_put(), safeguard the MUM_START
   done in error_return() to clear the SSF_NORET_VIA_MUMTSTART bit with a comment as to why this is
   okay to do so. Also, in dbg assert that we never encounter this bit set (cleanup should have
   been done in gvcst_put()).

3) Fix a longstanding typo by renaming SSF_NORET_VIA_MUMTSTART to SFF_NORET_VIA_MUMTSTART.
   SFF is to indicate this is a StackFrameFlag. SSF is a typo.