[#353] Nix GVINCRISOLATION error message as there seems to be no reason why simultaneous use of $INCREMENT and NOISOLATION should be disallowed (!360)
GVINCRISOLATION error documentation states $INCREMENT is currently not supported for globals that have
NOISOLATION turned on.

But because of how the global variable "in_gvcst_incr" gets reset to FALSE in gvcst_put() in between multiple
calls to gvcst_put2(), there is an inconsistency in when this error is issued. A test case documented
at https://gitlab.com/YottaDB/DB/YDB/issues/353 demonstrates this inconsistency.

While trying fix this inconsistency and issue GVINCRISOLATION error in all cases of NOISOLATION + $INCREMENT
usages, I started wondering why at all this error is necessary.

TP restarts for globals that have NOISOLATION turned on are avoided by redoing the SET operations at commit
time based on the current state of the leaf level block (which might have seen concurrent updates since the
SET command happened inside the TP transaction). This is done by the recompute_upd_array() function in
tp_tend.c.

If a $INCREMENT is done on a global node, it is treated as a SET operation where the current value of the
node is fetched and the increment is applied to it to get the desired post-increment-value and do a SET
of the node to the post-increment-value.

But in case the global node has $INCREMENT done on it while NOISOLATION is also turned ON, I suspect that
the original thinking behind the current implementation of this error message might be that the
recompute_upd_array() function needs to do an extra step of figuring out the current node value and
increment that value by the second parameter of the $INCREMENT operation while inside the commit logic
where it currently does not have access to this information.

But come to think of it, NOISOLATION being turned ON implies no other process can concurrently change this
particular node that the current process is attempting a $INCREMENT on. So we are guaranteed that if ever
recompute_upd_array() is invoked in tp_tend(), the node value would not have concurrently changed in between
the $INCREMENT and the TCOMMIT. This means it is okay to allow $INCREMENT and NOISOLATION usages together.

Removing the error will also remove the spanning-node inconsistency highlighted in the #353 test case.