[#354] Fix INVSTRLEN error to return correct string length and max length details in all cases
The INVSTRLEN error was issued in 3 places in the simpleAPI. 2 in sr_unix/libyottadb_int.h
and 1 in sr_unix/sapi_return_subscr_nodes.c. Out of these 3 usages, only 1 usage in
sr_unix/libyottadb_int.h provided the parameters needed by this error message (string length
and max length). The 2 other usages did not provide this detail which meant $ZSTATUS after
an INVSTRLEN error would contain values of 0 for string length and max length which is
not accurate. These 2 usages are now fixed to pass in the appropriate parameters so
$ZSTATUS is accurate.