[#352] Fix triggers/gtm8389 subtest failure due to a newly introduced incorrect assert (!358)
In this subtest, a spanning node set occurs and a trigger gets invoked inside the spanning set.
Once the trigger invocation is done, the SFF_NORET_VIA_MUMTSTART bit is reset in "gtm_trigger_fini".
And therefore the recently introduced assert (as part of #352 prior code commit) that the
SFF_NORET_VIA_MUMTSTART bit we added in "op_tstart" as part of the spanning node set be there at
"op_tcommit"time fails. This makes me think that manipulating this bit at "op_tstart" and "op_tcommit"
time for implicit TP in gvcst_put() is unnecessary and does not achieve any purpose. But I am
not 100% sure about that so don't want to delete its usage altogether in gvcst_put.c. Therefore
modified the assert to account for the fact that this bit might have been unset before we come to
op_tcommit. This addresses the test failure.