[#348] GTMASSERT2 fatal error during OPEN of a SOCKET that was closed after a TPTIMEOUT error during READ
See https://gitlab.com/YottaDB/DB/YDB/issues/348 for background/details.

The crux of the issue is that a TPTIMEOUT error (which causes "outofband" to be set to "tptimeout")
triggers the same actions that a job interrupt (MUPIP INTRPT) does in case of a SOCKET READ even though
the former does not restart the READ (since it triggers error processing) whereas the latter does
restart the read once it is done with executing $ZINTERRUPT M code.

This seems to be an issue that was missed out in GT.M V6.1-000 as part of GTM-7835 (see below release note).
There were 3 socket modules that could set dsocketptr->mupintr to TRUE (implying a socket action was
interrupted by a MUPIP INTRPT and needs to be restarted later). They were
	1) iosocket_wait.c,
	2) iosocket_connect.c
	3) iosocket_readfl.c

The first 2 of them were fixed to handle restartable issues in V6.1-000 but the last one was not.
Fixing iosocket_readfl.c just like the other two modules (to use OUTOFBAND_RESTARTABLE macro
to check if the "outofband" action is restartable or not and only set mupintr if it is restartable)
automatically fixes #348.