[#359] MUPIP INTRPT handling should save/restore transfer table entries so prior actions like ZSTEP continue to work after the job interrupt is handled
MUPIP INTRPT is handled by changing the transfer table entries in jobinterrupt_set() that way
at the next logical boundary (i.e. M line start etc.) the M code in $ZINTERRUPT is executed.
And the transfer table entry is reset in xfer_reset_handlers() to a pre-job-interrupt state.

It is this reset that is not correct. Since we do not save the current state of the transfer
table before changing it in jobinterrupt_set(), the reset that happens in xfer_reset_handlers()
is not correct in all cases. An example is if a ZSTEP action is in effect. This means once
the job interrupt is handled, any active ZSTEP action stops being active abruptly. To the user
this would be unfriendly as they do not control when the job interrupt happens but the tracing
they were running using ZSTEP suddenly stops working.

In this case, it seems more user friendly to save/restore the transfer table as part of
jobinterrupt_set() and xfer_reset_handlers(). This is now done using 6 global variables
defined in gbldefs.c.