Fix build issues on CYGWIN due to #include "rtnhdr.h" usages (should be #include <rtnhdr.h> instead)
The reason for why the <> syntax is needed for rtnhdr.h is captured in mdef.h in a comment block.

There were a few more modules that are part of multiple sr_* directories. The full list is below.
Note that sr_i386 is not included in the below since that is an unsupported platform for YottaDB.

   ./sr_unix/obj_filesp.h
   ./sr_unix_nsb/obj_filesp.h

   ./sr_port/opcode_def.h
   ./sr_unix_nsb/opcode_def.h

   ./sr_unix/rtnhdr.h
   ./sr_unix_nsb/rtnhdr.h

   ./sr_unix/zbreaksp.h
   ./sr_x86_64/zbreaksp.h
   ./sr_armv7l/zbreaksp.h

All the rest were already using <>. It was only rtnhdr.h that had "" usages in some places.
All of those are now fixed to use <>. All <> usages have now a comment indicating why this is necessary.

And unnecessary <rtnhdr.h> usages were removed.