[#348] Fix occasional v52000/C9G04002779 subtest failure due to a regression in a prior #348 commit
The #348 prior commit had removed a local variable "outofband_terminate" in iosocket_readfl.c and
instead replaced it with the global variable "outofband". But this is incorrect in case the only
string read in this invocation of the function is a delimiter and an outofband event (job interrupt)
simultaneously happens which would cause the delimiter to be ignored when saving the current
state of the read (to handle the job interrupt) resulting in data loss on the READ side of the socket.
Using the "outofband_terminate" local variable guarantees we do the save of the current read state
only if we have not yet seen a terminator and we see an outofband.