Revert "[#359] MUPIP INTRPT handling should save/restore transfer table entries so prior actions like ZSTEP continue to work after the job interrupt is handled"
This reverts commit dc0b865f20303bee0ce10473ec980b8697b0a882.

The reason is that jobinterrupt_set() (which changes entries in the transfer table) is interrupt code
that can be called at arbitrary points in time, including if interrupted code is actually in the process
of changing the transfer table entries itself (for example while compiling a $ZSTEP action like the
C-stack below indicates).

 #0  jobinterrupt_set (dummy_val=0) at sr_port/jobinterrupt_event.c:92
 #1  0x00007ffff715a67a in xfer_set_handlers (event_type=1, set_fn=0x7ffff71bbe86 <jobinterrupt_set>, param_val=0) at sr_port/deferred_events.c:209
 #2  0x00007ffff71bbe70 in jobinterrupt_event (sig=10, info=0x7fffffff8d30, context=0x7fffffff8c00) at sr_port/jobinterrupt_event.c:65
 #3  <signal handler called>
 #4  op_zst_break () at sr_port/op_zst_break.c:38
 #5  0x00007ffff6ececcf in stkok16 () at sr_x86_64/op_bkpt.s:93

In this case, if we save the current transfer table (to later restore in xfer_reset_handlers()),
we could be restoring an inconsistent transfer table later which might not be a safe thing to do.
For example, if 5 transfer table entries are in the process of being changed and the SIGUSR1 interrupt
occurs after 2 entries have been changed but the remaining 3 have not, then we would save (and in turn
later restore) a transfer table state that is inconsistent.

The previous code always reset the transfer table to hardcoded normal values which effectively lose the
ZSTEP action but that was still consistent in itself. So it is considered an unsafe change and undone.

As for the original issue of #359, I found a workaround and that is to redo the ZBREAK/$ZSTEP action
while inside the $ZINTERRUPT code. That ensures the ZSTEP action continues to be active even after the
$ZINTERRUPT. That is considered good enough for now.