[StaticScan] sr_unix/bin_load.c : Fix uninitialized variable "val" (introduced in GT.M V6.3-001 as part of GTM-7857)
In testing with pro and dbg builds, we were not able to reproduce any issue. So val was correctly
initialized to NULL in those cases. But it is not guaranteed. If it is a non-zero value, then
the mupip load process could even SIG-11. But since this is very unlikely to see in practice, it
is not considered a critical issue.