[StaticScan] sr_port/restrict.c : Cannot use SIZEOF(errstr) in SNPRINTF in append_filter() as errstr is a pointer (not an array) when passed in
Using SIZEOF on a "char *" pointer will cause the SNPRINTF to operate as if only 8 bytes are
available in errstr[] when actually it has space for MAX_FN_LEN + 1. The solution is to
make sure the SIZEOF(errstr) is done in the macro (which sees errstr as an array) where the
append_filter() call is done.