Fix debug-build-only occasional v53003/D9I10002703 subtest failure
Till date, we have noticed two failures of the mentioned subtest with the following signature.

%YDB-F-ASSERT, Assert failed in sr_unix/backup_buffer_flush.c line 97 for expression (EACCES == sbufh_p->backup_errno)

In both cases, backup_errno was ENOENT i.e. the temporary file was not present.
And that was because MUPIP BACKUP had cleaned up the temporary file as part of halting.
And it had halted with the following error.

> cat v53003_1_3/D9I10002703/SHELL/SHELL_MUPIP_BACKUP.log
.
.
Error doing unalias -a; cd v53003_1_3/D9I10002703 && /bin/cp --sparse=always mumps.dat v53003_1_3/D9I10002703/bakdir_SHELL/DEFAULT_29142__0.tmp
Error removing temp dir unalias -a; /bin/rm -r bakdir_SHELL/DEFAULT_29142__0.tmp.
%YDB-E-MUNOFINISH, MUPIP unable to finish all requested actions

This is because we do a SYSTEM() call inside mupip backup and that failed due to SHELL env var
being set to a huge value.

If it so happens that the background updates (started by the subtest) update a block while this
backup was trying to set things up (between mupip_backup() and mubfilcpy()), then it is possible
for the background update to invoke backup_buffer_flush() after its commit (in t_end/tp_tend)
and if the backup process had cleaned up the temporary file by then, this assert failure is very
much explainable.

The fix is to add an ENOENT clause as an || condition in the assert to allow for this condition.