[#357] SIGINT/SIGQUIT (kill -2 or kill -3) sent to a mumps process can sometimes be ignored/lost
gtm_system_internal() is a function that gets invoked when YottaDB need to spawn a process
(for example the ZSYSTEM command, scripts invoked by the $ydb_procstuckexec env var etc.)
and wait for the spawned process to return before resuming execution.

This currently ignores SIGINT/SIGQUIT and blocks SIGCHLD in order to mirror the system()
function call. But this means any SIGINT/SIGQUIT signals (kill -2 or kill -3) sent to the
mumps/mupip/etc. process, while it has spawned a child process through gtm_system_internal(),
will be permanently lost. This is not desirable.

This is now fixed by BLOCKing SIGQUIT/SIGINT for the duration of the spawn. This way any
such signals sent while the child is still running will be handled by the parent once the
child returns/exits. The signal handling is deferred instead of being lost permanently.

Also, a white-box case for WBTEST_BADEXEC_OP_ZEDIT was moved here from op_zedit.c as the
latter has been changed to invoke gtm_system_internal() in a different change (#360). This
is needed by the v62001/gtm8094zedproc subtest.