[#360] $ZEDITOR does not reflect exit status of the last ZEDIT invocation
ZEDIT command forks the editor and waits for it to complete. This logic is very similar to
that in gtm_system_internal() (gtm_system.c) so most of this logic is now instead replaced
with a call to that function.

And $ZEDITOR is now properly maintained based on the exit status of the invocation.
Previously it was non-zero (errno) only if wait() returned -1.  Whereas now it records the
exit status of the child in the most common case (when waitpid() returns something other
than -1). This is usually a positive number >=0 and < 256. In case the waitpid() returns -1
a SYSCALL error is now issued with errno included so the user gets the error detail.

Remove white box case for WBTEST_BADEXEC_OP_ZEDIT as that is now moved to gtm_system_internal().
This is needed by the v62001/gtm8094zedproc subtest.