Add -Wvla to identify uses of variable length arrays; Fix two existing usages of vla to not use vla
Variable Length Arrays are a convenience thing but are discouraged for a variety of reasons
(stack exhaustion, linear overflow, jumps over guard pages etc.). So add -Wvla to the build script
of YottaDB to identify such usages and fix them as soon as they are identified.

There were 2 usages of these in the current YottaDB code both of which came from recent GT.M versions.

1) sr_port/cert_blk.c : This usage was introduced in GT.M V6.3-003.
2) sr_unix/gtm_stdio.c : This usage was introduced in GT.M V6.3-005.

These are now fixed to not use Variable Length Arrays.