[#346] MUPIP INTEG, DSE DUMP and MUMPS should not infinite loop in case of INVSPECREC error; MUPIP INTEG should produce more user-friendly INVSPECREC error
The crux of the issue is in get_spec() where if the 1st byte of the 4-byte collation header
in the directory tree happens to be the special value of 2 (LAST_TYPE_DEFINED), then
it went into an infinite loop because spec_len[] array was defined only for 0 and 1 indices
but we would also access spec_len[2], because the check was (*ptr > LAST_TYPE_DEFINED) instead
of (*ptr >= LAST_TYPE_DEFINED), which meant that since spec_len[2] would most likely be 0,
the post-increment in the for loop would end up incrementing ptr by 0 which caused an infinite loop.

While a simple fix would be to replace ">" with ">=" in the (*ptr > LAST_TYPE_DEFINED) check,
it is not a complete fix because the current layout of get_spec() meant that any time we see
a 4-byte collation record where the 1st byte does not have the value 1 (COLL_SPEC), we would
issue an INVSPECREC error, be it in MUMPS or MUPIP INTEG or DSE DUMP.

In the case of MUMPS or DSE DUMP, it is not desirable to issue an error in this case. It would
be better to continue as if no collation was defined for this global. This is now fixed.

In the case of MUPIP INTEG, an error is definitely desirable but an error with more context is
needed. The INVSPECREC error just says there is a bad collation record header. It does not
indicate which block # and record # the issue exists. This is now fixed too.

As part of these changes, the logic in get_spec() was greatly simplified. And that is to just
look for the first 4-byte collation record (no need of a for loop of going through potentially
many 4-byte records). And simplify all callers of this function which had similar code all of
which is now moved into a new GET_GVT_COLL_INFO macro defined in spec_type.h. This meant that
get_spec.c and get_spec.h could be completely nixed/removed.

Also the INVSPECREC error is now only issued from MUPIP INTEG. And it has a !AD parameter because
of that (all MUPIP INTEG error messages have a !AD to indicate the block#/record# context leading
to the error).