sr_unix/iott_rdone.c : Fix assert (dbg-only code) which relied on uninitialized variable "timed"; Enhance iott_readfl.c to have similar asserts like in iott_rdone.c
The r120/readcmdrecallhist subtest (which was recently enhanced to test MUPIP INTRPT of READs)
failed the following assert.

%YDB-F-ASSERT, Assert failed in sr_unix/iott_rdone.c line 124 for expression (timed == tt_state->timed)

This is because "timed" was uninitialized at the point of the assert. This is now fixed by
using initialized variables to derive the value of "timed" at the point of the assert.