[#361] Fix rare case of REPLINSTNOHIST error from Receiver Server on restart after first A->P connection
See https://gitlab.com/YottaDB/DB/YDB/issues/361 for a description of the issue.

Basically, it is possible for jnlpool->jnlpool_ctl->strm_seqno[] to be 0 even in the
second A->P connection in case the receiver server was shutdown BEFORE a history record
got added by the update process in the first A->P connection. In that case, a restart of
the receiver server would see a strm_jnl_seqno of 0 and that would later cause a REPLINSTNOHIST
error (or assert failure in debug builds).

This is fixed by setting strm_jnl_seqno to 1 in case it is 0 in sr_unix/gtmrecv_process.c.
This is already done for another if/else block so that logic is moved outside the various
if/else codepaths with asserts added to any code paths where we are guaranteed the strm_jnl_seqno
will be non-zero.

Because of this change, a pre-existing assert in sr_unix/repl_instance.c is no longer true.
So it is relaxed a bit by allowing a condition where the jnlpool_ctl->strm_seqno[] seqno is 0
but the histinfo->strm_seqno is 1 (the exception that is now added in sr_unix/gtmrecv_process.c).