[#363] Compiler returns incorrect results in cases where it should issue NUMOFLOW
NUMOFLOWs are quietly ignored during s2n in compile time, and this wasn't being detected by the
numeric optimizations. As a result, the numeric optimization would continue without the numeric
representation of a given string value being filled out. In many cases, this meant that a value
of '0' would be used, resulting in incorrect behavior for boolean expressions.

In debug builds, the following assert failed in some cases.

%YDB-F-ASSERT, Assert failed in sr_port/ex_tail.c line 88 for expression (MV_NM & v0->mvtype)