[#364] Source Server shutdown command should give indication of whether it did not delete the journal pool ipcs even if the instance is frozen
In GT.M V6.3-000A, a change was done to return prematurely from repl_ipc_cleanup() if the instance
is frozen (i.e. IS_REPL_INST_FROZEN is TRUE). This meant that we did not fall through to code that
would have printed the "Not deleting jnlpool ipcs" message. To keep the output consistent even
if we return prematurely due to an instance freeze, a "Not deleting jnlpool ipcs" message is logged
with an reason/additional-detail being "Instance is frozen".