[#362] Avoid data differences between primary and secondary in case of kill -9 on secondary side
First some background on the issue
----------------------------------
When we ran the recently enabled manually_start/dual_fail2_no_ipcrm2 subtest, we noticed a failure
only on the ARM platform with the following diff.

> cat dual_fail2_no_ipcrm2.diff
53c53
< DATABASE EXTRACT PASSED
---
> TEST-E-FAILED: DATABASE EXTRACTs on PRIMARY and SECONDARY are DIFFERENT.

The databases on primary and secondary side were different at the end of the test. Below is the
actual data diff.

> cat pri_21_51_23.glo_sec_21_51_23.glo_spanreg_glodiff
7679c7679
< ^dntp(0,47902994)="somejunk"
---
> ^dntp(0,47902994)="-254.34"

This turned out to be a portable issue in YottaDB i.e. is an issue even on Linux/x86_64 builds.

The journal extract on the primary and secondary side show the following records corresponding
to the data that is different.

Primary
  TSTART \64912,78256\36354\3272188576\32446\0\4147\0\0
  TSET   \64912,78256\36354\3272188576\32446\0\4147\0\0\4\1\^dntp(0,47902994)="somejunk"
  TCOM   \64912,78256\36354\3699605187\32446\0\4147\0\0\4\BA
  SET    \64912,78256\36355\1971748481\32446\0\4148\0\0\0\0\^dntp(0,47902994)="-254.34"

Secondary
  TSTART \64912,78293\36354\2396333726\32276\0\4147\0\0
  TSET   \64912,78293\36354\2396333726\32276\0\4147\0\0\4\1\^dntp(0,47902994)="somejunk"
  TCOM   \64912,78293\36354\2523728995\32276\0\4147\0\0\4\BA
  NULL   \64912,78293\36355\2306412254\32276\0\4148\0\0
  ALIGN  \64912,78293\0\3780385109\32264\0

Notice how there were two SETs of ^dntp on the primary side, but there is only one SET on the
secondary side. The second SET on the primary side, which happened at jnlseqno of 4148 happened on
the secondary as a NULL record. This is possible because the test did a crash of the secondary side
where the update process was killed using kill -9 but the database shared memory was not removed (like
it would if this was a system crash). Note that since GT.M V6.3-002 (and YottaDB r1.10), as part of
GTM-8436 (http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V6.3-002_Release_Notes.html#GTM-8436),
journal records are now committed in two phases. If the update process was in the process of
committing seqno 4148 when it got killed, it is possible it had reserved space in the journal buffers
(done in phase1 of commit while holding the db critical section lock) but before it wrote the journal
records in phase2 of commit (when it does not hold the db critical section lock) it got killed.

In this state, later when the test did a MUPIP JOURNAL -ROLLBACK -BACKWARD on the secondary, it
found out database shared memory still lying around so tried to flush whatever it can out to disk
before starting the rollback. It found the reserved space for the SET record in the journal buffer
section but had no idea of the actual journal record content (since this is a different process
than the update process) and so filled that space with a NULL/ALIGN journal record combination
(the 0x50 space of the SET record was split as 0x30 for a NULL record and 0x20 for an ALIGN
record). And after running down the database shared memory in this fashion, it proceed with the
rollback which played forward this seqno=4148 with a NULL journal record effectively creating a
difference in data between the primary and secondary.

Although kill -9 is not supported, a rollback should have handled this better by recognizing
this NULL record was not a user-initiated NULL record but instead was a filler NULL record and so
identify this as a broken transaction and play this not onto the database but into the broken transaction
file. That would have rolled the secondary back to a journal sequence number that is before the
NULL record and so a future connection with the primary would have refetched the user-generated
journal records for this sequence number from the primary thereby avoiding the data discrepancy issue.

The proposed fix is to record in the NULL record (used in replication and non-replication environments
respectively to fill up reserved space) the fact that these are auto-generated records that way
any journal recover/rollback can safely treat these as broken transactions and avoid the data
discrepancy issue altogether.

Fixes
-----
* sr_port/jnl.h : The struct_jrec_null structure now has a "salvaged" bit that indicates this is an
  auto-generated NULL record if the bit is set to 1 and a user-generated NULL record if it is set to 0.

* sr_port/jnl_phase2_salvage.c and sr_port/repl_phase2_salvage.c : The "salvaged" bit is set to 1 to
  indicate this is an auto-generated NULL record.

* sr_port/mur_back_process.c : The crux of the change is here. If we find a NULL journal record with
  the "salvaged" bit set and this is a fetchresync rollback (i.e. rollback on the receiver side of
  replication), we treat this auto-generated update as potentially having a different journal record
  for the same seqno on the corresponding source side of replication. Therefore we treat this as a
  broken transaction.

* sr_port/jnl.h and sr_port/repl_filter.h : Because of the format change in the NULL journal record, the
  journal format has to be bumped. And because the NULL record is replicated, the replication filter
  format also has to be bumped. But because GT.M versions can concurrently do bump of the formats
  to the exact same values YottaDB bumps them to, there needs to be a buffer built in between to
  accommodate GT.M version bumps to these formats. Therefore, the current journal format of GDSJNL27
  has been bumped to GDSJNL44 as part of this change leaving a buffer of 16 format changes from the
  GT.M side in the future to be incorporated into YottaDB without any issues. The format number 44 is
  what gets stored in the journal file. The GDSJNL part of GDSJNL44 is only displayed by
  MUPIP JOURNAL -SHOW and so that is changed from GDSJNL to YDBJNL.

* sr_port/muprec.h : The journal extract format and detailed extract format both have to be bumped.
  To avoid issues with concurrent GT.M changes to the extract format, the extract format label has been
  changed from GDS to YDB. So the new extract format is YDBJEX08 and the new detailed extract format is
  YDBJDX09.

* sr_port/repl_filter.h and sr_port/repl_filter.c : Since we only need to support older versions as old
  as 5 years old, the earliest GT.M version supported has been bumped up from V5.0-000 to V6.0-000. This
  meant we need to only upto an older format of V22 (instead of V17) and so a lot of V17, V19 and V21
  related code has been removed. And new V44 related code has been added. And various other code cleanup
  was done and comments added as appropriate.

* sr_port/mur_jnl_ext.c : The NULL record journal extract now includes the "salvaged" bit both in the
  regular and detailed extracts.

* sr_port/updproc.c and sr_port/mur_output_record.c : Now propagate/preserve the "salvaged" bit from
  the source NULL record while generating a new NULL record. This ensures that a NULL record which has
  the "salvaged" bit set on one source instance continues to have the "salvaged" bit set on another
  receiving instance in replication. This way if A->B->C replication is in place, a NULL record with
  "salvaged" bit set in B shows up with the bit set on C too so in case C does a fetchresync rollback
  with A, this NULL record is extracted as a broken transaction on C even though it became a salvaged
  record on B.

* Misc cleanup : repl_tr_endian_convert() was defined in gtmrecv_process.c and gtmsource_process.c but
  both were completely different functions. To avoid confusion these have been renamed with a _rcvr
  and _src suffix (i.e. repl_tr_endian_convert_rcvr() and repl_tr_endian_convert_src()).

* Messages like REPLNOXENDIAN are no longer possible due to the oldest version bump to V22 and so have
  been nixed.